add arg require in Dockerfile, finish Makefile

+++ Makefile
@@ -14,32 +14,33 @@ rikka-test:
 
 IMAGE_NAME = 7sdream/rikka
 OLD_VERSION = $(shell docker images | sed -n 's:$(IMAGE_NAME) \+\([0-9.]\+\).*:\1:gp')
+HAS_LATEST = $(shell docker images | sed -n 's:$(IMAGE_NAME) \+\(latest\).*:\1:gp')
 NEW_VERSION = $(shell sed -n 's:\t*Version = "\([0-9.]\+\).*":\1:p' api/consts.go)
 GIT_COMMIT = $(strip $(shell git rev-parse --short HEAD))
 
 version:
-	@echo "Will delete $(OLD_VERSION)"
-	@echo "Will build $(NEW_VERSION)"
+	@echo "Will delete $(OLD_VERSION) $(HAS_LATEST)"
+	@echo "Will build $(NEW_VERSION) latest"
 
-ifeq ($(OLD_VERSION),$(NEW_VERSION))
-confirm:
-	$(error No newer version than $(OLD_VERSION))
-else
 confirm:
 	@bash -c "read -s -n 1 -p 'Press any key to continue, Ctrl+C to stop'"
-endif
 
 delete:
+ifneq ($(OLD_VERSION),)
 	docker rmi $(IMAGE_NAME):$(OLD_VERSION)
+endif
+ifneq ($(HAS_LATEST),)
 	docker rmi $(IMAGE_NAME):latest
+endif
+	-docker rmi $(IMAGE_NAME):$(NEW_VERSION)
 
-build: version confirm clean delete 
-	docker build \
+build: version confirm delete clean
+	@docker build \
 		--build-arg VCS_URL=$(shell git config --get remote.origin.url) \
   		--build-arg VCS_REF=$(GIT_COMMIT) \
   		--build-arg BUILD_DATE=$(shell date -u +"%Y-%m-%dT%H:%M:%SZ") \
 		-t $(IMAGE_NAME):$(NEW_VERSION) .
-	docker tag -f $(IMAGE_NAME):$(NEW_VERSION) $(IMAGE_NAME):latest
+	docker tag $(IMAGE_NAME):$(NEW_VERSION) $(IMAGE_NAME):latest
 
 push: build 
 	dockler push $(IMAGE_NAME)
